Move configs to root
